+++ config/wdio/selenium.config.js
@@ -21,6 +21,11 @@ const chromeConfig = {
 const firefoxConfig = {
   browserName: 'firefox',
   maxInstances: 1,
+  'moz:firefoxOptions': {
+    prefs: {
+      'dom.disable_beforeunload': false,
+    },
+  },
 };
 
 const ieConfig = {